Improve the downloading of files when sharing or exporting sounds
Use the TableObject download status to get the state of the file and to show the current state of the download when the file is already downloading. Also add the DownloadStatus enum in the Sound class and GetAudioFileDownloadStatus in Sound and GetSoundFileDownloadStatus in FileManager to get the download status of the table object file.