Save the sound order in Dictionaries and add LoadSelectedCategory to load sounds
Instead of reading the sound order from the database when selecting a category, load the sound order when the app is started and save the order in the CustomSoundOrder dictionaries to sort the sounds fast when selecting a category.
Also fix a bug which occured when selecting a category while all sounds were loading by adding LoadSelectedCategory method, which is called by LoadAllSoundsAsync, LoadSoundsByCategoryAsync and LoadSoundsByNameAsync. It makes sure that all sounds are loading once, and when the user selects a category while the sounds are loading it calls itself again.